Give the scheduler its own module
The scheduler was the second half of cli.py: deciding what the schedule owes
and running it has nothing to do with reading a command line, and the module
docstring already described the two as separate jobs.

It could not leave before now. It calls the plugin for a snapshot, a send, a
purge, a sync and a removal, those five lived in cli.py, and cli.py has to
import the scheduler to start its thread. Now that api.py holds the calls, the
move is a move: scheduler.py imports api.py and schedule.py, cli.py imports
scheduler.py, and nothing points back.

The thread stays with cli.py, which starts it beside the server and stops it on
a signal, since that is also when the socket goes. The set of replications
under way follows run_replicate, its only reader.

Nothing is changed on the way, not a line of the functions that moved.
